2022-2023/Groupe_7/innovafrica_api/requirements.txt
apturl==0.5.2
asgiref==3.5.2
attrs==19.3.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
cachetools==4.0.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
cryptography==38.0.3
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.14.0
django-filter==22.1
django-multiselectfield==0.1.12
django-prometheus==2.2.0
djangorestframework==3.14.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
GDAL==3.0.4
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
html5lib==1.0.1
httplib2==0.14.0
idna==3.4
importlib-metadata==1.5.0
ipython_genutils==0.2.0
Jinja2==2.10.1
jsonschema==3.2.0
jupyter-core==4.6.3
keyring==18.0.1
kiwisolver==1.0.1
labelImg==1.8.6
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.1.2
monotonic==1.5
more-itertools==4.2.0
mysqlclient==2.1.1
nbformat==5.0.4
netifaces==0.10.4
numpy==1.24.3
oauth2client==4.1.2
oauthlib==3.2.2
olefile==0.46
OWSLib==0.19.1
packaging==20.3
pandas==2.0.2
paramiko==2.6.0
pexpect==4.6.0
Pillow==9.3.0
plotly==4.4.1
prometheus-client==0.15.0
protobuf==3.6.1
psycopg2==2.8.4
pyamp==1.2
pyanp==0.3.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
pycups==1.9.73
PyDrive==1.3.1
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyproj==2.5.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.6
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
sqlparse==0.4.3
systemd-python==234
tomli==2.0.1
traitlets==4.3.3
ttystatus==0.38
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.26.12
usb-creator==0.3.7
wadllib==1.3.3
webencodings==0.5.1
xkit==0.0.0
zipp==1.0.0
